+++ libraries/base/changelog.md
@@ -13,6 +13,9 @@
 
   * Add `Data.ByteArray` module, providing a boxed `ByteArray#`.
 
+  * `fromEnum` for `Natural` now throws an error for any number that cannot be
+    repesented exactly by an `Int` (#20291).
+
 ## 4.16.0.0 *TBA*
 
   * Add a `Typeable` constraint to `fromStaticPtr` in the class `GHC.StaticPtr.IsStatic`.